apps/smart-routing/requirements.txt
Flask==2.3.3
pandas==2.2.2
python-dateutil==2.9.0
xgboost==2.0.3
scikit-learn==1.5.0
numpy==1.26.4